cross-account-region-data-pipeline/requirements-dev.txt
attrs==21.4.0
aws-cdk-lib==2.15.0
aws-cdk.aws-codestar-alpha==2.15.0a0
aws-cdk.cloud-assembly-schema==2.15.0
aws-lambda-powertools==1.25.2
aws-xray-sdk==2.9.0
boto3==1.21.13
botocore==1.24.13
cattrs==1.10.0
constructs==10.0.81
crhelper==2.0.10
data==0.4
decorator==5.1.1
fastjsonschema==2.15.3
funcsigs==1.0.2
future==0.18.2
jmespath==0.10.0
jsii==1.54.0
marshmallow==3.14.1
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.8
wrapt==1.13.3
awswrangler~=2.15.0
pandas~=1.3.5
setuptools~=47.1.0
